docs/tools/requirements.txt
aiohttp
beautifulsoup4
click
mkdocs
mkdocs-material
mkdocs-material-extensions
mypy
pycodestyle
pydocstyle
pylint
pymdown-extensions
PyYAML
python-dateutil
python-dotenv
requests
tqdm
pydantic
types-requests
